Add files via upload
Here is the Mathematica file for the program where I would like polarization sums (replacing two polarizations with a Kronecekr Delta) then define a rule to impose the Kronecker delta and change the indices properly. In the end, I need to make a tensor contraction as much as possible since the final expression must be very compact due to cancelations and simplification which one can not see before doing this final step. Note that the tensors K and F, do not have any specific symmetries therefore whenever we flip their indices we should transpose them.